update supporting ts fully
